fixed target langauge generation workflow
