+++ src/js/actions/TargetLanguageActions.js
@@ -3,12 +3,6 @@ import path from 'path-extra';
 
 // actions
 import * as ResourcesActions from './ResourcesActions';
-// helpers
-import * as USFMHelpers from '../helpers/usfmHelpers';
-import { getBibleIndex } from '../helpers/ResourcesHelpers';
-import * as UsfmFileConversionHelpers from '../helpers/FileConversionHelpers/UsfmFileConversionHelpers';
-import {saveTargetBible} from "../helpers/Import/importHelpers";
-import {processChapter} from "../helpers/Import/tStudioHelpers";
 
 /**
  * @description Loads a target language bible chapter from file system.
@@ -37,63 +31,5 @@ export function loadTargetLanguageChapter(chapterNumber) {
       dispatch(ResourcesActions.addNewBible(resourceId, bibleId, bibleData));
     }
   });
-}
-
-/**
- * Processes USFM file to create tC project
- * @param usfmFilePath
- * @param projectPath
- * @param manifest
- */
-export function generateTargetBibleFromUSFMPath(usfmFilePath, projectPath, manifest) {
-  try {
-    let usfmFile;
-    let parsedUSFM;
-    const filename = path.basename(usfmFilePath);
-    usfmFilePath = fs.existsSync(usfmFilePath) ? usfmFilePath : path.join(projectPath, filename);
-    if (fs.existsSync(usfmFilePath)) {
-      usfmFile = fs.readFileSync(usfmFilePath).toString();
-      parsedUSFM = USFMHelpers.getParsedUSFM(usfmFile);
-    } else {
-      console.warn('USFM not found');
-      return;
-    }
-    const { chapters } = parsedUSFM;
-    let targetBible = {};
-    Object.keys(chapters).forEach((chapterNumber)=>{
-      const chapterObject = chapters[chapterNumber];
-      targetBible[chapterNumber] = {};
-      Object.keys(chapterObject).forEach((verseNumber)=>{
-        const verseData = chapterObject[verseNumber];
-        targetBible[chapterNumber][verseNumber] = UsfmFileConversionHelpers.getUsfmForVerseContent(verseData);
-      });
-    });
-    saveTargetBible(projectPath, manifest, targetBible);
-  } catch (error) {
-    console.warn('USFM not found');
-    console.log(error);
-  }
-}
-
-/**
- * Processes tStudio project files to create tC project
- * @param projectPath
- * @param manifest
- */
-export function generateTargetBibleFromTstudioProjectPath(projectPath, manifest) {
-  let bookData = {};
-  // get the bibleIndex to get the list of expected chapters
-  const bibleIndex = getBibleIndex('en', 'ult');
-  if(!bibleIndex[manifest.project.id]) {
-    console.warn(`Invalid book key ${manifest.project.id}. Expected a book of the Bible.`);
-    return;
-  }
-  let header = '';
-  const chapters = Object.keys(bibleIndex[manifest.project.id]);
-  chapters.unshift( 'front', '0' ); // check for front matter folders
-  chapters.forEach(chapter => {
-    header = processChapter(chapter, projectPath, header, bookData);
-  });
-  saveTargetBible(projectPath, manifest, bookData, header);
 }
